+++ src/Notify.ts
@@ -2,6 +2,9 @@ import { NotificationManager } from "./notification-manager";
 
 const globalManager = new NotificationManager();
 
+/**
+ * @deprecated Use `snackbar()` instead
+ */
 const notify: (notification: {
     message: string;
     duration?: number;
@@ -16,4 +19,28 @@ const notify: (notification: {
     classes?: Array<string> | string;
 }) => void = globalManager.notify.bind(globalManager);
 
-export { NotificationManager, notify };
+const snackbar: (notification: {
+    message: string;
+    duration?: number;
+    closeable?: boolean;
+    buttons?: Array<{
+        label: string;
+        callback: Function;
+        ariaLabel?: string;
+        classes?: Array<string> | string;
+    }>;
+    force?: boolean;
+    classes?: Array<string> | string;
+}) => void = globalManager.snackbar.bind(globalManager);
+
+const toast: (notification: {
+    title: string;
+    message: string;
+    closeable?: boolean;
+    icon?: string;
+    duration?: number;
+    classes?: string[];
+    element?: HTMLElement;
+}) => void = globalManager.toast.bind(globalManager);
+
+export { NotificationManager, notify, toast, snackbar };